requirements/base.txt
black==19.3b0
bleach==3.1.0
defusedxml==0.6.0
django>=1.11.19,<2.0
django-autocomplete-light>=3.2.10,<3.3
django-contrib-comments==1.9.1
django-countries==5.5
django-crispy-forms==1.7.2
django-cron==0.5.1
django-durationfield==0.5.5
django-filter>=1.1.0,<2.0
django-modeltranslation==0.13.3
django-premailer
django-request==1.5.5
django-reversion==3.0.4
django-taggit==0.23.0
djangorestframework==3.9.4
djmail==1.1.0
docutils==0.15.2
factory_boy>=2.10.0,<3.0
mwoauth==0.3.7
pillow==6.2.1
pypandoc==1.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
wheel==0.33.6
whitenoise==4.1.4
